fix to 32 bit ewma filter

+++ src/sapphireos/util.c
@@ -256,8 +256,8 @@ uint16_t util_u16_ewma( uint16_t new, uint16_t old, uint8_t ratio ){
 
 uint32_t util_u32_ewma( uint32_t new, uint32_t old, uint8_t ratio ){
 
-    uint32_t temp = ( ( (uint32_t)ratio * new ) / 256 ) +  
-                    ( ( (uint32_t)( 256 - ratio ) * old ) / 256 );
+    uint32_t temp = ( ( (uint64_t)ratio * new ) / 256 ) +  
+                    ( ( (uint64_t)( 256 - ratio ) * old ) / 256 );
 
     // check if filter is unchanging
     if( temp == old ){
